Reorganize codebase into library and binary structure
Split duplihere into a clean library/binary architecture for better maintainability:

Library structure (src/lib.rs):
- Extract all core duplicate detection functionality to library
- Make functions public for external testing and reuse
- Implement proper error handling throughout
- Organize code into logical modules and sections

Binary structure (src/main.rs):
- Reduce to focused CLI interface (82 lines vs 1,580+ lines)
- Use library functions for all core functionality
- Clean separation of concerns between CLI and algorithms

Test organization (tests/integration_tests.rs):
- Move all 26 comprehensive unit tests to external test file
- Enable proper integration testing of library functions
- Maintain full test coverage of core algorithms
- Follow standard Rust testing conventions

Benefits:
- Better code maintainability and organization
- Improved testability with external test access
- Standard Rust project structure
- Cleaner separation between library and CLI
- Enables library reuse by other projects

All functionality preserved with significantly improved code organization.
